Used jellyfish for levenstein and made parallel
Jellyfish's implementation for levenshtein distance is at least twice
as fast as the previous package and meets all our needs. To make it even
faster, implemented with multiprocessing to make use of all system
threads during the computation of the distance.

TODO: we have yet to make the distance useful in our context. We
currently return the lowest distance, or most similar query. This is
dependent on the length of a phrase as well (longer phrases may need
more substitutions and changes). So, we need to come up with a better
way to come up with similarity based on maybe average length of dataset
or by weighting the length of the phrases as well compared to the
dataset length.